### 2022-08-26 #4713 @editorialbot
[PRE REVIEW]: OPSDN: an enhanced SDN simulation framework for OPNET Modeler
**Submitting author:** @ZacharyJia (<a href="http://orcid.org/0000-0002-7144-4448">Zequn Jia</a>)
**Repository:** https://github.com/ZacharyJia/opsdn
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.1
