+++ paper.md
@@ -30,7 +30,12 @@ bibliography: paper.bib
 ---
 
 # Summary
-Traditional SDN simulators/emulators like OPNET, Mininet [@mininet], Estinet[@estinet] and MaxiNet[@maxinet], cannot provide acceptable simulation capabilities in large-scale and/or high-traffic network scenarios due to different event scheduling system. We propose OPSDN, a simulation framework that unifies the evetlet[@eventlet] event scheduling of the existing Ryu controller framework with the traditional OPNET DES (Discrete Event Simulation) engine. In OPSDN, the SDN controller events are also triggered by the DES engine, ensuring the consistency of the simulation progress between the two sides, thus obtaining more accurate simulation results. Additionally, OPSDN provides Ryu-like APIs and can reuse part of the Ryu libraries. Our simulation results show that the OPSDN controller can achieve the same control effect as the typical Ryu controller, and avoid unexpected TCP retransmissions due to the inconsistent simulation speed. We hope OPSDN will help researchers achieve more accessible large-scale SDN network simulations. Detailed introduction and evaluation are included in @jia_opsdn_2022.
+Simulation tools are one of the essential softwares in the research of software-defined networking (SDN). Tradational SDN simulation softwares includes testbeds like Mininet[@mininet] and network simulators with SDN module support such as OPNET. But both kinds of these simulation tools suffer from distortion when in large scale topology or high-traffic scenarios. Mininet usually support network devices no more than 4096 and its CPU and bandwidth are limited by a single server. OPNET needs an interface to connect simualtion and external SDN controller, which requires the simulation speed is strictly the same with real world time. But simulation speed is generally slower than realtime in large scale topologies due to hardware limition.
+
+To solve the problem above, we propose OPSDN, a simulation framework for OPNET that allows researchers to condcut *large-scale* SDN simulations in laptop-level hardwares. Instead of an external controller, OPSDN integrate the controller into the simulation to accurately sync the time of switches and the controller. It also offers controller APIs similar to the popular SDN controller framework Ryu[@ryu]. Researcher could use familiar APIs to develop and port controller applications.
+Our simulation results show that the OPSDN controller can achieve the same control capability as the Ryu controller, and avoid unexpected TCP retransmissions introduced by the inconsistent simulation speed. We hope OPSDN will help researchers achieve more accessible large-scale SDN network simulations. Detailed introduction and evaluation are included in @jia_opsdn_2022.
+
+
 
 # Statement of need
 OPSDN provides researchers with a network simulation framework that allows them to conduct large-scale SDN network simulations on laptop-level hardware and leverage the accurate simulation modelling capabilities provided by OPNET. Therefore, the main impacts of this software are as follows.
