+++ paper.bib
@@ -1,30 +1,3 @@
-
-@misc{eventlet,
-  title = {Eventlet},
-  year = {2022},
-  month = jun,
-  journal = {Eventlet Networking Library},
-  howpublished = {https://eventlet.net/},
-  langid = {english}
-}
-
-@misc{libpcap,
-  title = {Libpcap},
-  year = {2022},
-  month = jun,
-  journal = {TCPDump \& Libpcap},
-  howpublished = {https://www.tcpdump.org/},
-  langid = {english}
-}
-
-@misc{opnet_benchmark,
-  title = {{{OPNET SITL}} Benchmark},
-  year = {2022},
-  month = jun,
-  howpublished = {https://github.com/ZacharyJia/OPNET-SITL-Benchmark},
-  langid = {english}
-}
-
 @misc{ryu,
   title = {Ryu},
   year = {2022},
@@ -48,96 +21,6 @@
   isbn = {978-1-4503-0409-2},
   keywords = {emulation,open-flow,rapid prototyping,software defined networking,virtualization}
 }
-
-@article{7464256,
-  title = {{{SDN-SPS}}: {{Semi-physical}} Simulation for Software-Defined Networks},
-  author = {Hu, Long and Wang, Junfeng and Song, Enmin and Ksentini, Adlen and Hossain, Mu{\d h}ammad Anwar and Rawashdeh, Majdi},
-  year = {2016},
-  journal = {IEEE Sensors Journal},
-  volume = {16},
-  number = {20},
-  pages = {7355--7363},
-  doi = {10.1109/JSEN.2016.2561969}
-}
-
-@phdthesis{hyden_time_2003,
-  title = {Time Dilation: {{Decreasing}} Time to Decision with Discrete -Event Simulation},
-  author = {Hyden, Paul D.},
-  year = {2003},
-  journal = {ProQuest Dissertations and Theses},
-  abstract = {Time dilation is proposed as a general method for selecting the best performing system from a discrete search space. As proposed by Schruben (1997), time dilation describes a method for managing the simultaneous execution of discrete event simulation models. Instead of fixing the time scale across all concurrently executing models, Schruben contracts the time scales of the most interesting models and dilates the time scales of less interesting models. The net result is that more computer time is spent simulating interesting models, improving decision making in the process. In this work, this concept is motivated and developed in depth, and its effectiveness is demonstrated in realistic simulation experiments. Several of the key computations have changed and the overall mechanism for implementation is completely new. Before giving the details of the fundamental time dilation algorithm, several background issues are discussed. A broad survey of the simulation optimization literature is provided. Further discussion establishes concretely why some models should receive more simulation work and how time dilation can satisfy those needs. A tool for measuring the basic experimental unit of simulation work, extended event graphs, is introduced. Additionally, extended event graphs enable techniques for reducing memory requirements that may be strained by simultaneous execution of multiple models. Time dilation and its attendant key issues are discussed and a general algorithm is presented. Implementations of time dilation for two examples and computational evidence showing significant improvements of time dilation over current approaches are presented.},
-  isbn = {978-0-493-96094-4},
-  langid = {english},
-  keywords = {0796:Operations research,Applied sciences,Decision-making,Discrete-event simulation,Operations research,Time dilation},
-  file = {C\:\\Users\\jia19\\Zotero\\storage\\9G86A5EQ\\Hyden - 2003 - Time dilation Decreasing time to decision with di.pdf}
-}
-
[…]
