### 2022-09-08 #4713 @ZacharyJia
> @ZacharyJia I am going to ping the editorial board to see if this seems in scope for us to review. This will take 1-2 weeks.
>
> In the meantime, please rewrite the beginning of your paper for people who aren't familiar with your area of expertise. It has a lot of jargon and acronyms, including those that aren't written out.

Hi @kthyng , thanks and I have updated the summary of this paper.
