+++ paper.md
@@ -32,7 +32,7 @@ bibliography: paper.bib
 # Summary
 Simulation tools are one of the essential softwares in the research of software-defined networking (SDN). Tradational SDN simulation softwares includes testbeds like Mininet[@mininet] and network simulators with SDN module support such as OPNET. But both kinds of these simulation tools suffer from distortion when in large scale topology or high-traffic scenarios. Mininet usually support network devices no more than 4096 and its CPU and bandwidth are limited by a single server. OPNET needs an interface to connect simualtion and external SDN controller, which requires the simulation speed is strictly the same with real world time. But simulation speed is generally slower than realtime in large scale topologies due to hardware limition.
 
-To solve the problem above, we propose OPSDN, a simulation framework for OPNET that allows researchers to condcut *large-scale* SDN simulations in laptop-level hardwares. Instead of an external controller, OPSDN integrate the controller into the simulation to accurately sync the time of switches and the controller. It also offers controller APIs similar to the popular SDN controller framework Ryu[@ryu]. Researcher could use familiar APIs to develop and port controller applications.
+To solve the problem above, we propose OPSDN, a simulation framework for OPNET that allows researchers to conduct *large-scale* SDN simulations in laptop-level hardwares. Instead of an external controller, OPSDN integrate the controller into the simulation to accurately sync the time of switches and the controller. It also offers controller APIs similar to the popular SDN controller framework Ryu[@ryu]. Researcher could use familiar APIs to develop and port controller applications.
 Our simulation results show that the OPSDN controller can achieve the same control capability as the Ryu controller, and avoid unexpected TCP retransmissions introduced by the inconsistent simulation speed. We hope OPSDN will help researchers achieve more accessible large-scale SDN network simulations. Detailed introduction and evaluation are included in @jia_opsdn_2022.
 
 
