### 2022-09-28 #4713 @danielskatz
Can you now make suggestions for potential reviewers by please mentioning them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).
