### 2022-10-26 #4815 @Qingfengmufeng
OK

***@***.***

From: Daniel S. Katz
Date: 2022-10-25 20:20
To: openjournals/joss-reviews
CC: Qingfengmufeng; Mention
Subject: Re: [openjournals/joss-reviews] [REVIEW]: OPSDN: an enhanced SDN simulation framework for OPNET Modeler (Issue #4815)
@Qingfengmufeng - can you try that again? enter @editorialbot generate my checklist as a new comment here
