### 2023-02-13 #4815 @pradeeban
@danielskatz Happy to help. Please assign me the paper as a reviewer.

Before I start, one thing I noticed is the README.

The module-level READMEs are in Chinese.

https://github.com/ZacharyJia/opsdn/tree/master/src/dll-bridge

https://github.com/ZacharyJia/opsdn/tree/master/src/models

@ZacharyJia, Is there a chance to include the English translations as part of the README?

That will help me understand and appreciate the context and contribution of the framework better.
