### 2023-02-13 ZacharyJia/opsdn#1 @pradeeban
Installation with pip/pip3 fails
$ pip3 --version
pip 22.2 from /Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages/pip (python 3.7)
Pradeebans-MacBook-Pro:~ pradeeban$ pip3 install opsdnpy
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -rllib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution - (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -rllib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution - (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
Collecting opsdnpy
  Using cached opsdnpy-0.2.1-py3-none-any.whl (4.8 kB)
  Using cached opsdnpy-0.2-py3-none-any.whl (4.6 kB)
  Using cached opsdnpy-0.1.3-py3-none-any.whl (4.5 kB)
ERROR: Cannot install opsdnpy==0.1.3, opsdnpy==0.2 and opsdnpy==0.2.1 because these package versions have conflicting dependencies.

The conflict is caused by:
    opsdnpy 0.2.1 depends on networkx==2.8.2
    opsdnpy 0.2 depends on networkx==2.8.2
    opsdnpy 0.1.3 depends on networkx==2.8.2

To fix this you could try to:
1. loosen the range of package versions you've specified
2. remove package versions to allow pip attempt to solve the dependency conflict

ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/topics/dependency-resolution/#dealing-with-dependency-conflicts
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -rllib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution - (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -rllib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution - (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -rllib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution - (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)

[notice] A new release of pip available: 22.2 -> 23.0
[notice] To update, run: pip install --upgrade pip

-----------------or----------------------------

$ pip --version
pip 22.2 from /Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages/pip (python 3.7)
Pradeebans-MacBook-Pro:~ pradeeban$ pip install opsdnpy
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -rllib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution - (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -rllib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution - (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
Collecting opsdnpy
  Using cached opsdnpy-0.2.1-py3-none-any.whl (4.8 kB)
Requirement already satisfied: ryu==4.34 in /Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages (from opsdnpy) (4.34)
  Using cached opsdnpy-0.2-py3-none-any.whl (4.6 kB)
  Using cached opsdnpy-0.1.3-py3-none-any.whl (4.5 kB)
ERROR: Cannot install opsdnpy==0.1.3, opsdnpy==0.2 and opsdnpy==0.2.1 because these package versions have conflicting dependencies.

The conflict is caused by:
    opsdnpy 0.2.1 depends on networkx==2.8.2
    opsdnpy 0.2 depends on networkx==2.8.2
    opsdnpy 0.1.3 depends on networkx==2.8.2

To fix this you could try to:
1. loosen the range of package versions you've specified
2. remove package versions to allow pip attempt to solve the dependency conflict

ERROR: ResolutionImpossible: for help visit https://pip.pypa.io/en/latest/topics/dependency-resolution/#dealing-with-dependency-conflicts
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -rllib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution - (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -rllib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution - (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -rllib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution -llib3 (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)
WARNING: Ignoring invalid distribution - (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/site-packages)

[notice] A new release of pip available: 22.2 -> 23.0
[notice] To update, run: pip install --upgrade pip
