### 2023-02-13 ZacharyJia/opsdn#2 @pradeeban
[Documentation] READMEs in the src/ folder lack English descriptions
The module-level READMEs in the src folder are in Chinese.

https://github.com/ZacharyJia/opsdn/tree/master/src/dll-bridge

https://github.com/ZacharyJia/opsdn/tree/master/src/models

Is there a chance to include the English translations as part of the README.

That will help us understand and appreciate the context and contribution of the framework better.
