### 2023-02-13 #4815 @pradeeban
Thanks @danielskatz.

@ZacharyJia, I am attempting to install the package, following the pip install command as listed in the README. It fails. I have reported this as a bug in https://github.com/ZacharyJia/opsdn/issues/1

Can you please attend to it, so that I can proceed with the rest of the review?

I have also opened a separate issue on README lacking English translations at https://github.com/ZacharyJia/opsdn/issues/2 for ease of reference/tracking.
