### 2023-02-14 #4815 @ZacharyJia
> Thanks @danielskatz.
>
> @ZacharyJia, I am attempting to install the package, following the pip install command as listed in the README. It fails. I have reported this as a bug in [ZacharyJia/opsdn#1](https://github.com/ZacharyJia/opsdn/issues/1)
>
> Can you please attend to it, so that I can proceed with the rest of the review?
>
> I have also opened a separate issue on README lacking English translations at [ZacharyJia/opsdn#2](https://github.com/ZacharyJia/opsdn/issues/2) for ease of reference/tracking.

Hi @pradeeban ,
Thanks for your help! I will follow up these issues and try to fix them.
