### 2023-02-14 ZacharyJia/opsdn#1 @ZacharyJia
Hi @pradeeban ,
1. I think the first problem is that you are using python 3.7, while the prebuilt binary is for Python3.8. If you would like to use the prebuilt library, I would suggest you to install a python3.8. Or you have to build the dll-bridge yourself.

2. Could you please check if you have already installed a different version of networkx before? It might be the reason of the dependency problem from pip.

And I just noticed that you are using a Mac? This repo does not provide Mac support because OPNET will not run on Mac. And I also don't have access to a linux version of OPNET, so linux version is also not involved.
In a nutshell, **only windows version** is supported now.
