### 2023-02-14 ZacharyJia/opsdn#1 @pradeeban
Thanks, @ZacharyJia for the clarification.

I will close this issue since it is mentioned in the README clearly about the OS and Python requirements.

I do not have access to Windows. But I will proceed with the review.
