### 2023-02-14 ZacharyJia/opsdn#1 @pradeeban
Tested `pip install opsdnpy` on Python 3.8 on Ubuntu. It worked.
