### 2023-02-14 #4815 @pradeeban
@ZacharyJia I am reading the paper now and leaving some comments along the way.

The statement "Tradational SDN simulation softwares includes testbeds like Mininet" has a few issues.

1. Mininet is an emulator and not a simulator. Although many people tend to use "emulation" and "simulation" interchangeably, the differences are significant in networking. Mininet and Maxinet are network emulators. NS-3 and CloudSimSDN are network simulators. You may need to include both simulation and emulation examples, as in, "Tradational SDN simulators includes NS-3 and emulators include Mininet."

The paper
https://www.researchgate.net/publication/322472564_A_Feature_Based_Comparison_of_SDN_Emulation_And_Simulation_Tools gives some more examples of the simulators vs. emulators.

If you want to avoid introducing emulators in your paper, use NS-3 as an example for simulators instead of Mininet.

2. The typo in the word "Traditional" above.
