### 2023-02-14 #4815 @pradeeban
Along the same comments above, a performance comparison of Mininet vs. OPNET will be unfair. Since Mininet is an emulator, it will indeed be resource-heavy. (Mininet creates processes to represent each host and switch, whereas simulators do not do that.)

"Mininet usually support network devices no more than 4096 and its CPU and bandwidth are limited by a single server."

If you need to continue the comparisons, these definitions and the emulation vs. simulation discussions should be presented beforehand. Some revisions to the "Summary" will be necessary for accuracy and fairness in these comparisons.
