### 2023-02-14 #4815 @pradeeban
@ZacharyJia Please make sure to define the acronyms the first time they appear in the paper. SD-MANET and DES lack such definitions in the paper.
