### 2023-02-14 #4815 @pradeeban
@ZacharyJia The last sentence mentions NS3 and OMNET++, which are indeed the simulators (and not Mininet-like emulators). They should be mentioned early on in the summary for accuracy and clarity. Such a comparison against NS-3 will make it clearer where your contribution stands, compared against the simulators (NS-3, OMNET++, ...) and emulators (Mininet, Maxinet, ...)

I am aware of the space limitations in JOSS. But addressing the accuracy of the claims will make the paper more convincing.
