### 2023-02-14 #4815 @pradeeban
@ZacharyJia, if you can address my major comments in the paper and produce a revised version, I will continue the review on the revised paper.

1. The emulator vs simulator clarification early on in the summary and throughout the paper. Mininet should be mentioned as emulator, and performance claims should be ideally made against simulators (such as NS-3), for fairness. Functional claims can be made against simulators and emulators. Overall, I propose some writeup enhancements for accuracy and completeness. Please feel free to get back to me if any of my comments are unclear.

2. Please do a spellcheck and grammar check using a software such as Grammarly. I spotted some typos.

3. Please define acronyms and more citations (especially to the network simulators) as appropriate.

4. Module-level documentation should be improved. You have 3 modules. 2 of them have README in Chinese. One does not have README.

5. Please note the current state of "Review checklist for @pradeeban" is filled with my current review of the paper and the open source software.
