+++ paper.bib
@@ -28,3 +28,21 @@
   year = {2022},
   url = {https://github.com/ZacharyJia/opsdn/blob/master/opsdn-technical-report.pdf}
 }
+
+@misc{ns3,
+  title = {ns-3: a discrete-event network simulator for internet systems},
+  year = {2022},
+  month = jun,
+  journal = {NS3 Network Simulator},
+  howpublished = {https://www.nsnam.org/},
+  langid = {english}
+}
+
+@inproceedings{omnetpp,
+  title={THE OMNET++ DISCRETE EVENT SIMULATION SYSTEM},
+  author={Andr{\'a}s Varga},
+  year={2001},
+  booktitle = {Proceedings of the European Simulation Multiconference (ESM2001)},
+  address = {Czech Republic},
+  articleno = {9},
+}
