+++ paper.bib
@@ -4,7 +4,8 @@
   month = jun,
   journal = {Ryu SDN Framework},
   howpublished = {https://ryu-sdn.org/},
-  langid = {english}
+  langid = {english},
+  note = {Accessed on 06.01.2022}
 }
 
 @inproceedings{mininet,
@@ -35,7 +36,8 @@
   month = jun,
   journal = {NS3 Network Simulator},
   howpublished = {https://www.nsnam.org/},
-  langid = {english}
+  langid = {english},
+  note = {Accessed on 02.22.2023}
 }
 
 @inproceedings{omnetpp,
