+++ paper.md
@@ -30,7 +30,7 @@ bibliography: paper.bib
 ---
 
 # Summary
-Network testing environments are essential softwares in the research of software-defined networking (SDN). Traditional SDN testing tools include emulators like Mininet[@mininet] and simulators with SDN module support such as OPNET, NS3[@ns3] and OMNET++[@omnetpp]. However, both kinds of these tools suffer from distortion when in large-scale topology or high-traffic scenarios. The emulation of network devices typically requires heavy resources, making it difficult for emulators to support large-scale network simulations. For example, Mininet usually supports network devices no more than 4096 and its CPU and bandwidth are limited by a single server. Simulators can usually support larger-scale networks by reducing the simulation speed. However, OPNET needs an interface to connect the simulation to an external SDN controller, which requires the simulation speed to be strictly the same as real-world time. But simulation speed is generally slower than real-time in large-scale topologies due to hardware limitations.
+Network testing environments are essential softwares in the research of software-defined networking (SDN)[@zavrak2017feature]. Traditional SDN testing tools include emulators like Mininet[@mininet] and simulators with SDN module support such as OPNET, NS3[@ns3] and OMNET++[@omnetpp]. However, both kinds of these tools suffer from distortion when in large-scale topology or high-traffic scenarios. The emulation of network devices typically requires heavy resources, making it difficult for emulators to support large-scale network simulations. For example, Mininet usually supports network devices no more than 4096 and its CPU and bandwidth are limited by a single server. Simulators can usually support larger-scale networks by reducing the simulation speed. However, OPNET needs an interface to connect the simulation to an external SDN controller, which requires the simulation speed to be strictly the same as real-world time. But simulation speed is generally slower than real-time in large-scale topologies due to hardware limitations.
 
 To solve the problem above, we propose OPSDN, a simulation framework for OPNET that allows researchers to conduct *large-scale* SDN simulations in laptop-level hardware. Instead of an external controller, OPSDN integrates the controller into the simulation to accurately sync the time of switches and the controller. It also offers APIs similar to the popular SDN controller framework Ryu[@ryu]. Researchers could use familiar APIs to develop and port controller applications.
 Our simulation results show that the OPSDN controller can achieve the same control capability as the Ryu controller, and avoid unexpected TCP retransmissions introduced by the inconsistent simulation speed. We hope OPSDN will help researchers achieve more accessible large-scale SDN network simulations. Detailed introduction and evaluation are included in @jia_opsdn_2022.
