+++ paper.bib
@@ -48,3 +48,11 @@
   address = {Czech Republic},
   articleno = {9},
 }
+
+@inproceedings{zavrak2017feature,
+  title={A feature-based comparison of SDN emulation and simulation tools},
+  author={Zavrak, S and Iskefiyeli, M},
+  booktitle={Proc. Int. Conf. Eng. Technol.(ICENTE)},
+  pages={214--217},
+  year={2017}
+}
