### 2023-02-22 6524a4b
Bug fix with paper and docs updated (#3)
* document update

* fix the problem that we have to use fixed controller ip and port

* fix typo

* paper revision
