+++ paper.bib
@@ -4,7 +4,8 @@
   month = jun,
   journal = {Ryu SDN Framework},
   howpublished = {https://ryu-sdn.org/},
-  langid = {english}
+  langid = {english},
+  note = {Accessed on 06.01.2022}
 }
 
 @inproceedings{mininet,
@@ -27,4 +28,31 @@
   author = {Jia, Zequn and Sun, Yantao and Liu, Qiang},
   year = {2022},
   url = {https://github.com/ZacharyJia/opsdn/blob/master/opsdn-technical-report.pdf}
+}
+
+@misc{ns3,
+  title = {ns-3: a discrete-event network simulator for internet systems},
+  year = {2022},
+  month = jun,
+  journal = {NS3 Network Simulator},
+  howpublished = {https://www.nsnam.org/},
+  langid = {english},
+  note = {Accessed on 02.22.2023}
+}
+
+@inproceedings{omnetpp,
+  title={THE OMNET++ DISCRETE EVENT SIMULATION SYSTEM},
+  author={Andr{\'a}s Varga},
+  year={2001},
+  booktitle = {Proceedings of the European Simulation Multiconference (ESM2001)},
+  address = {Czech Republic},
+  articleno = {9},
+}
+
+@inproceedings{zavrak2017feature,
+  title={A feature-based comparison of SDN emulation and simulation tools},
+  author={Zavrak, S and Iskefiyeli, M},
+  booktitle={Proc. Int. Conf. Eng. Technol.(ICENTE)},
+  pages={214--217},
+  year={2017}
 }
