+++ src/models/sdn_controller_agent.pr.cpp
@@ -1,450 +1,0 @@
-/* Process model C++ form file: sdn_controller_agent.pr.cpp */
-/* Portions of this file copyright 2017 by Riverbed Technology. All rights reserved. */
-
-
-/*
- =========================== NOTE ==========================
- This file is automatically generated from sdn_controller_agent.pr.m
- during a process model compilation.
-
- Do NOT manually edit this file.
- Manual edits will be lost during the next compilation.
- =========================== NOTE ==========================
-*/
-
-
-
-/* This variable carries the header into the object file */
-const char sdn_controller_agent_pr_cpp [] = "MIL_3_Tfile_Hdr_ 1860 30A op_runsim_opt 7 62A5D1DE 62A5D1DE 1 Zachary-1080 lab_user 0 0 none none 0 0 none 0 0 0 0 0 0 0 0 4e52 0                                                                                                                                                                                                                                                                                                                                                                                              ";
-#include <string.h>
-
-
-
-/* OPNET system definitions */
-#include <opnet.h>
-
-
-
-/* Header Block */
-
-#include "sdn_controller_agent/header_block.h"
-
-/* End of Header Block */
-
-#if !defined (VOSD_NO_FIN)
-#undef	BIN
-#undef	BOUT
-#define	BIN		FIN_LOCAL_FIELD(_op_last_line_passed) = __LINE__ - _op_block_origin;
-#define	BOUT	BIN
-#define	BINIT	FIN_LOCAL_FIELD(_op_last_line_passed) = 0; _op_block_origin = __LINE__;
-#else
-#define	BINIT
-#endif /* #if !defined (VOSD_NO_FIN) */
-
-
-
-/* State variable definitions */
-class sdn_controller_agent_state
-	{
-	private:
-		/* Internal state tracking for FSM */
-		FSM_SYS_STATE
-
-	public:
-		sdn_controller_agent_state (void);
-
-		/* Destructor contains Termination Block */
-		~sdn_controller_agent_state (void);
-
-		/* State Variables */
-		Objid	                  		my_node_objid                                   ;
-		ApiT_Tcp_App_Handle	    		tcp_app_handle                                  ;
-		std::map<int, ApiT_Tcp_App_Handle>			tcp_app_handle_map                              ;
-
-		/* FSM code */
-		void sdn_controller_agent (OP_SIM_CONTEXT_ARG_OPT);
-		/* Diagnostic Block */
-		void _op_sdn_controller_agent_diag (OP_SIM_CONTEXT_ARG_OPT);
-
-#if defined (VOSD_NEW_BAD_ALLOC)
-		void * operator new (size_t) throw (VOSD_BAD_ALLOC);
-#else
-		void * operator new (size_t);
-#endif
-		void operator delete (void *);
-
-		/* Memory management */
-		static VosT_Obtype obtype;
-	};
-
-VosT_Obtype sdn_controller_agent_state::obtype = (VosT_Obtype)OPC_NIL;
-
-#define my_node_objid           		op_sv_ptr->my_node_objid
[…]
