+++ src/opsdnpy/README.md
@@ -1,0 +1,9 @@
+# opsdnpy
+
+This folder contains the python part of OPSDN.
+
+
+## Structures
+
+- `examples`: Here we provide some examples of opsdnpy, including a controller class, a simple_switch app, a topology_discovery app and a fattree_fw app.
+- `package`: The core opsdnpy package in Pypi.
