### 2023-02-22 ZacharyJia/opsdn#5 @pradeeban
Community guidelines
Do you have clear guidelines for third parties wishing to
1) Contribute to the software
2) Report issues or problems with the software
3) Seek support?

A contributing.md or a wiki page addressing the above should help.
