### 2023-02-22 ZacharyJia/opsdn#6 @pradeeban
Automated tests
Do/Can you have automated tests or sufficiently documented manual steps to verify the functionality?

Please let me know if you have documentation somewhere for the manual steps (or point me to the automated tests).

It will be good to document such steps, if no automated tests available.
