### 2023-02-28 #4815 @pradeeban
Hi @danielskatz, I am still awaiting @ZacharyJia's responses to https://github.com/ZacharyJia/opsdn/issues/6 and https://github.com/ZacharyJia/opsdn/issues/5 before ticking them.

When I checked, I couldn't locate them. If @ZacharyJia updates those issues, I can proceed with these final two checks of the review.
