### 2023-03-07 ZacharyJia/opsdn#6 @danielskatz
Note that the JOSS criterion for tests is in https://joss.readthedocs.io/en/latest/review_criteria.html#tests
