+++ docs/functionality-check-manual.md
@@ -1,0 +1,158 @@
+# Functionality Check manual
+
+This manual aims to provide a user guide for testing the features provided by this software.
+
+<!-- omit in toc -->
+## Table of Contents
+
+- [Prerequisite Requirements](#prerequisite-requirements)
+- [Installation](#installation)
+- [Basic Workflow Test](#basic-workflow-test)
+- [Common API Test](#common-api-test)
+  - [`@controller_app` and `@set_ev_handler`](#controller_app-and-set_ev_handler)
+  - [`opsdn.sim_time()` and `OpTimer.set_op_timer()`](#opsdnsim_time-and-optimerset_op_timer)
+
+## Prerequisite Requirements
+1. OS version: Windows 7 or newer
+1. Compiler: Microsoft Visual Studio 2010 (Recommended by Riverbed Modeler)
+1. OPNET(Riverbed) Modeler 18.6.1
+1. Python 3.8 with `Python38.dll` included in the `PATH` environment variables
+
+## Installation
+### Steps:
+1. Install `opsdnpy`：
+```bash
+pip install opsdnpy
+```
+
+2. Download the prebuilt binary file from this link (https://github.com/ZacharyJia/opsdn/releases/download/v0.2.1/opsdn-prebuilt-vs2019-x64-v0.2.1.zip) and unzip it.
+3. Create a new folder named "opsdn" and place the files `library.h`, `opsdn.dll`, and `opsdn.lib` into it. Then add the full path of this folder to the `include`, `path`, and `lib` environment variables.
+4. Copy all remaining files in the prebuilt folders to the `op_models` directory in the user's home directory.
+
+## Basic Workflow Test
+### Steps:
+1. Create a new opnet project and scenario.
+2. Add nodes (a sdn_controller_ethernet_wkstn_adv, a of_switch_eth4 and two ethernet_wkstn_adv) to your scenario, and connect them like the following figure.
+
+![simple_switch_topo](./images/simple_switch_topo.png)
+
+3. Save the example controller file [simple_switch.py](https://github.com/ZacharyJia/opsdn/blob/master/src/opsdnpy/examples/simple_switch.py) to your local folder.
+
+4. Right click the node and click [Edit Attributes] to configure node attributes. The configurations are as follows:
+- [controller]-[IP]-[IP Host Parameters]-[Interface Information]-[Address] => `192.168.100.1`
+- [controller]-[ControllerPy]-[Python Home] => Your Python3.8 home
+- [controller]-[ControllerPy]-[Script Path] => folder of simple_switch.py
+- [controller]-[ControllerPy]-[Entry Script] => `simple_switch`
+- [controller]-[ControllerPy]-[Entry Class] => `MyController`
+- [controller]-[ControllerPy]-[OpenFlow Port] => `998`
+- [switch]-[IP]-[IP Host Parameters]-[Interface Information]-[Address] => `192.168.100.2`
+- [switch]-[Port To Controller] => Usually P0, but you should check by right click the link between controller and switch.
+- [switch]-[Controller IP Address] => `192.168.100.1`
+- [switch]-[Controller TCP Port] => `998` (This must lower than 1024, or will raise some bug due to OPNET's TCP design)
+
+5. Add an `application demand` between node_1 and node_2, configure some traffic. For example:
+- [Request Parameters]-[Size] => exponential(256)
+- [Request Parameters]-[Rate] => exponential(3600)
+- [Transport Protocol] => TCP
+
+6. Right Click the scenario, click [Choose individual DES statistics]. Check the [Node statistics]-[Application Demand].
+
+7. Press Ctrl+R to run the simulation.
+
+8. After run the simulation, right click the scenario and choose [view results].
[…]
