### 2023-03-08 ZacharyJia/opsdn#6 @ZacharyJia
Hi @pradeeban and @danielskatz ,

Thanks for pointing this out. As it is really hard to make automated tests for OPNET, I added a [Functionality Check manual](https://github.com/ZacharyJia/opsdn/blob/master/docs/functionality-check-manual.md) for you to check OPSDN's features.

Hope this will help.
