### 2023-03-08 ZacharyJia/opsdn#6 @pradeeban
This documentation/manual is OK, as per the JOSS criterion for tests, especially considering the difficulty in creating automated tests for frameworks such as OPNET. I close this issue as fixed.
