### 2023-03-08 #4815 @ZacharyJia
> @ZacharyJia - at this point could you:
>
>     * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.
>
>     * [ ]  Archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
>
>     * [ ]  Check the archival deposit (e.g., in Zenodo) has the correct metadata. This includes the title (should match the paper title) and author list (make sure the list is correct and people who only made a small fix are not on it). You may also add the authors' ORCID.
>
>     * [ ]  Please list the DOI of the archived version here.
>
>
> I can then move forward with acceptance of the submission, which will involve a dry run, proofreading the paper at that point, and then either suggesting small changes or moving directly to acceptance and publication

Hi @danielskatz,

  Thanks for your help. I have finished these steps and here is the version tag and DOI:

Archived version tag:  `v0.2.2`
DOI in Zenodo:  `10.5281/zenodo.7708397`
