+++ paper.bib
@@ -31,7 +31,7 @@
 }
 
 @misc{ns3,
-  title = {ns-3: a discrete-event network simulator for internet systems},
+  title = {ns-3: a discrete-event network simulator for {I}nternet systems},
   year = {2022},
   month = jun,
   journal = {NS3 Network Simulator},
@@ -41,7 +41,7 @@
 }
 
 @inproceedings{omnetpp,
-  title={THE OMNET++ DISCRETE EVENT SIMULATION SYSTEM},
+  title={The OMNET++ Discrete Event Simulation System},
   author={Andr{\'a}s Varga},
   year={2001},
   booktitle = {Proceedings of the European Simulation Multiconference (ESM2001)},
